requirements.txt
numpy==1.19.2
Pillow==8.4.0
fastapi==0.105.0
uvicorn==0.24.0
scikit-learn==0.24.1